used double in total hours to return accura total hours in decimal

+++ src/main/java/com/tensura/dtrsystem/entity/AttendanceEntity.java
@@ -28,7 +28,7 @@ public class AttendanceEntity {
 
     private String timeOutImage;
 
-    private int totalHours;
+    private Double totalHours;
 
     @Column(precision = 10, scale = 7)
     private BigDecimal latitude;
